URM-F/UFM3- FederatedAgents/requirements.txt
# requirements.txt
torch
torchvision
Pillow
streamlit
matplotlib
opencv-python
